allow y-position to be a float

+++ lib/flappy_elixir/components/y_position.ex
@@ -3,5 +3,5 @@ defmodule FlappyElixir.Components.YPosition do
   Documentation for YPosition components.
   """
   use ECSx.Component,
-    value: :integer
+    value: :float
 end
